Only calculate path once until the destination is reached
