+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/attributes/goals/MoveToGoal.kt
@@ -1,9 +1,9 @@
 package de.gleex.pltcmd.game.engine.attributes.goals
 
 import de.gleex.pltcmd.game.engine.GameContext
+import de.gleex.pltcmd.game.engine.commands.MoveTo
 import de.gleex.pltcmd.game.engine.entities.types.ElementEntity
 import de.gleex.pltcmd.game.engine.entities.types.position
-import de.gleex.pltcmd.game.engine.systems.facets.MoveTo
 import de.gleex.pltcmd.model.world.coordinate.Coordinate
 import org.hexworks.amethyst.api.Command
 import org.hexworks.cobalt.datatypes.Maybe
