+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/systems/facets/PathFinding.kt
@@ -1,10 +1,12 @@
 package de.gleex.pltcmd.game.engine.systems.facets
 
 import de.gleex.pltcmd.game.engine.GameContext
+import de.gleex.pltcmd.game.engine.commands.MoveTo
+import de.gleex.pltcmd.game.engine.entities.types.destination
 import de.gleex.pltcmd.game.engine.entities.types.movementPath
 import de.gleex.pltcmd.game.engine.entities.types.position
 import de.gleex.pltcmd.model.world.coordinate.Coordinate
-import de.gleex.pltcmd.util.geometry.pointsOfLine
+import de.gleex.pltcmd.model.world.coordinate.CoordinatePath
 import org.hexworks.amethyst.api.Command
 import org.hexworks.amethyst.api.Consumed
 import org.hexworks.amethyst.api.Response
@@ -20,14 +22,24 @@ object PathFinding : BaseFacet<GameContext>() {
     override suspend fun executeCommand(command: Command<out EntityType, GameContext>): Response {
         return command.responseWhenCommandIs(MoveTo::class) {
             val (destination, _, entity) = it
-            val pathStack = Stack<Coordinate>()
+            log.info("Current destination: ${entity.destination}")
+            if(entity.movementPath.isNotEmpty()
+                    && entity.destination.filter { current -> current == destination }.isPresent) {
+                log.debug("Destination and path are already set.")
+                return@responseWhenCommandIs Consumed
+            }
             val currentPosition = entity.position.value
             log.debug("Finding path from $currentPosition to $destination...")
-            pointsOfLine(currentPosition.eastingFromLeft, currentPosition.northingFromBottom, destination.eastingFromLeft, destination.northingFromBottom) { x, y ->
-                pathStack.push(Coordinate(x, y))
-            }
-            pathStack.reverse()
+            val pathStack = Stack<Coordinate>()
+                    .apply {
+                        addAll(
+                            CoordinatePath
+                                .line(currentPosition, destination)
+                                .drop(1)
+                                .reversed())
+                    }
             log.debug("Setting path: $pathStack")
+            log.debug("Fist step: ${pathStack.peek()}")
             entity.movementPath = pathStack
             Consumed
         }
